requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
autopep8==2.3.1
billiard==4.2.0
celery==5.4.0
celery-progress==0.4
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==42.0.8
cssselect2==0.7.0
Deprecated==1.2.14
Django==5.0.6
django-appconf==1.0.6
django-backup-utils==0.2.7
django-celery-beat==2.6.0
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-dbbackup==4.1.0
django-filter==24.2
django-redis==5.4.0
django-select2==8.2.1
django-storages==1.14.3
django-tables2==2.7.0
django-timezone-field==6.1.0
factory_boy==3.3.1
Faker==28.0.0
gunicorn==22.0.0
html5lib==1.1
idna==3.7
kombu==5.3.7
lxml==5.2.2
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pycodestyle==2.12.0
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pyparsing==3.1.2
pypdf==4.2.0
PyPDF3==1.0.6
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.7
reportlab==4.2.2
requests==2.32.3
rules==3.5
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
xhtml2pdf==0.2.16
